requirements.txt
langchain
langchain-community
langchain-openai
langchain-experimental
streamlit
pandas
matplotlib
seaborn
pygwalker
jupyterlab
tiktoken
tabulate
python-dotenv
langchain-google-genai

# install langchain ollama for local llm deployment
langchain-ollama
openpyxl # to read .xlsx file